move DelEntry to entry_del.go
